renamed PermissionFragment to SetupFragment, references #6
