+++ app/src/main/java/de/vdvcount/app/ui/setup/SetupFragment.java
@@ -1,4 +1,4 @@
-package de.vdvcount.app.ui.permission;
+package de.vdvcount.app.ui.setup;
 
 import androidx.lifecycle.ViewModelProvider;
 
@@ -14,18 +14,18 @@ import android.view.ViewGroup;
 
 import de.vdvcount.app.R;
 
-public class PermissionFragment extends Fragment {
+public class SetupFragment extends Fragment {
 
-    private PermissionViewModel mViewModel;
+    private SetupViewModel mViewModel;
 
-    public static PermissionFragment newInstance() {
-        return new PermissionFragment();
+    public static SetupFragment newInstance() {
+        return new SetupFragment();
     }
 
     @Override
     public void onCreate(@Nullable Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        mViewModel = new ViewModelProvider(this).get(PermissionViewModel.class);
+        mViewModel = new ViewModelProvider(this).get(SetupViewModel.class);
         // TODO: Use the ViewModel
     }
