+++ app/src/main/res/navigation/navigation_graph.xml
@@ -3,11 +3,11 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:id="@+id/navigation_graph.xml"
-    app:startDestination="@id/permissionFragment">
+    app:startDestination="@id/setupFragment">
 
     <fragment
-        android:id="@+id/permissionFragment"
-        android:name="de.vdvcount.app.ui.permission.PermissionFragment"
+        android:id="@+id/setupFragment"
+        android:name="de.vdvcount.app.ui.setup.SetupFragment"
         android:label="fragment_permission"
         tools:layout="@layout/fragment_permission" />
 </navigation>
